% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baker}
\alias{baker}
\title{Citation data between social work journals for the 1985-86 period}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\usage{
data(baker)
}
\description{
This example consists of the citation data between social 
work journals for the 1985-86 period, collected and analyzed 
in Baker (1992)
}
\examples{
# data(baker)
# Transforming it to matrix format
# baker <- as.matrix(baker)
# putting zeros on the diagonal
# diag(baker) <- 0 
}
\references{
\enc{Baker, D. R.}{Baker, D. R.} (1992). A Structural Analysis of Social Work Journal Network: 1985-1986. Journal of Social Service Research, 15(3-4), 153-168. doi: 10.1300/J079v15n03_09
}
\keyword{datasets}
