% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canClu.R
\encoding{UTF-8}
\name{canClu}
\alias{canClu}
\alias{canCluUniqe}
\title{Create canonical partition and find unique canonical partitions in a list of partitions.}
\usage{
canClu(clu)

canCluUniqe(cluList)
}
\arguments{
\item{clu}{A partition - a vector or a list of vectors/partitions.}

\item{cluList}{A list of partitions(vectors).}
}
\value{
For function \code{canClu} - a canonical partition or a list of such partitions.
For function \code{canCluUniqe} - A list of unique canonical partitions.
}
\description{
It is used to convert any partition to a canonical partition. A canonical partition is a partition where the first unit is in cluster 1, the next unit that is not in cluster 1 in in cluster 2 and so on. So if we would take first appearances of clusters in the order they appear in the partition vector, we would get integers from 1 to the number of clusters.
}
\examples{
clu<-c(3,2,2,3,1,2)
canClu(clu)
}
\seealso{
\code{\link{clu}}
}
