% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeToZeroIndexed.R
\name{encodeToZeroIndexed}
\alias{encodeToZeroIndexed}
\title{Encode Integer Vector with Consecutive Values Starting from 0}
\usage{
encodeToZeroIndexed(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
An integer vector with values from 0 to (number of unique values - 1),
maintaining the original order structure.
}
\description{
This function maps an integer vector to a new vector where the unique values
are replaced with consecutive integers starting from 0.
}
\examples{
encodeToZeroIndexed(c(10, 20, 10, 30))
# Returns: 0 1 0 2

}
