% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandMat.R
\encoding{UTF-8}
\name{expandMat}
\alias{expandMat}
\title{Expands a square matrix by repeating each row/column the specified number of times.}
\usage{
expandMat(mat, nn)
}
\arguments{
\item{mat}{A square matrix to be exapanded}

\item{nn}{A vector of number of times each row/column must be repeated. Its length must match the number of rows/columns}
}
\value{
Sum of squared deviations from the mean using only valid (non NA) values.
}
\description{
Expands a square matrix by repeating each row/column the specified number of times.
}
\author{
\enc{Aleš Žiberna}{Ales Ziberna}
}
\keyword{manip}
