% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanRep.R
\encoding{UTF-8}
\name{nanRep}
\alias{nanRep}
\title{Replaces NaN values by the speficied values (0 by default)}
\usage{
nanRep(x, rep = 0)
}
\arguments{
\item{x}{A vector or similar where the NaNs are to be replaced.}

\item{rep}{A value that should replace the NaNs (0 by default).}
}
\value{
x with NaNs replaced by rep.
}
\description{
Replaces NaN values by the speficied values (0 by default)
}
\author{
\enc{Aleš Žiberna}{Ales Ziberna}
}
\keyword{manip}
