% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printBlocks.R
\encoding{UTF-8}
\name{printBlocks}
\alias{printBlocks}
\title{Niceprinting of the \code{blocks} parameter as used in \code{\link{optRandomParC}} and \code{\link{critFunC}}.}
\usage{
printBlocks(blocks)
}
\arguments{
\item{blocks}{\code{blocks} parameter as used in \code{\link{optRandomParC}} and \code{\link{critFunC}}.}
}
\value{
Used for side effects (printing)
}
\description{
Niceprinting of the \code{blocks} parameter as used in \code{\link{optRandomParC}} and \code{\link{critFunC}}.
}
\seealso{
\code{\link{optRandomParC}}, \code{\link{critFunC}}
}
\author{
\enc{Aleš, Žiberna}{Ales Ziberna}
}
\keyword{print}
