% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relInv.R
\encoding{UTF-8}
\name{relInv}
\alias{relInv}
\alias{relInv2}
\title{Functions for computing "relative inverse" (\code{x[1]/x}).}
\usage{
relInv(x)

relInv2(x)
}
\arguments{
\item{x}{A numeric vector. For \code{relInv} it should not contain 0s (while for  \code{relInv2} it can).}
}
\value{
A vector computed as \code{x[1]/x}. For \code{relInv2}, if the non-finite elements are replaced with 0s.
}
\description{
For a vector x, it computes \code{x[1]/x}. For \code{relInv2}, if certain elements of the result are not finite (e.g. if certain elements of x are 0), these elements are replaced with 0s.
}
\author{
\enc{Aleš Žiberna}{Ales Ziberna}
}
\keyword{manip}
