\name{ss}
\alias{ss}
\alias{ad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Sum of Squared deviations from the mean and sum of Absolute Deviations from the median}
\usage{
ss(x)
as(x)
}
\value{
Sum of Squared deviations from the mean or sum of Absolute Deviations from the median}
\author{Aleš Žiberna}
\keyword{univar}% at least one, from doc/KEYWORDS
